Update README.md
We are creating a user management system. We need to be able to add users to our database. A user has a username and a password. The password should not be easy to guess, so it should consist of one lowercase character, one uppercase character and one number. 

Here, a user has to enter his/her password and name. While the programme can handle exceptions, if the user enters an invalid password, it will simply ask the user to do it again!